Add spicy spinach chicken macaroni recipe.
A recipe is a json object with a name and, optionaly, an operation and
ingredients, and amount and unit. An operation is what is done to the
ingredients to create the named object. The ingredients field is a list
of recipies to wich the operation is done. Primitive ingredients have
amounts (numerical) and units (string).